fix(코드숨 2주차 과제1) 카운터 앱 만들고 분리하기 안녕하세요 트레이너님! 리뷰 요청드립니다! 항상 좋은 피드백 감사합니다!
