+++ src/components/ViewCount/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function CounterView({ count, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      Click me!(
+      {count}
+      )
+    </button>
+  );
+}
+
+export default CounterView;
